Migrate all handler scripts into handlers dir
